feat: add sonner for toast notifications and implement toast handling in the app

+++ src/actions/handle-note.ts
@@ -3,6 +3,9 @@ import 'server-only';
 import { parseWithZod } from "@conform-to/zod";
 import { noteActionSchema } from "./schema";
 import { prisma } from '@/utils/db.server';
+import { cookies } from 'next/headers';
+import { revalidatePath } from 'next/cache';
+import { redirect } from 'next/navigation';
 
 
 export async function handleNote(prevState: unknown, formData: FormData){
@@ -31,22 +34,47 @@ export async function handleNote(prevState: unknown, formData: FormData){
             break;
         case 'delete':
             // Handle deletion logic here
-
-            console.log('deleting note', submission.value.id);
-            await prisma.note.delete({
-                where: {
+            const note = await prisma.note.findFirst({
+                where:{
                     id: submission.value.id
                 }
             })
-            // revalidatePath('/notes');
-            console.log('note deleted');
-            return { status: 'success', message: 'this is fucking sick' } as const;
-            
+
+            if(!note){
+                return {
+                    status: 'error',
+                    message: 'Note not found',
+                    submission: submission.reply()
+                } as const;
+            }
+
+            await prisma.note.delete({
+                where: {
+                    id: note.id
+                }
+            })
+
+            const cookieStore = await cookies();
+            cookieStore.set({
+                name: 'n_toast',
+                value: JSON.stringify({
+                    description: 'Note deleted successfully',
+                    type: 'success'
+                }),
+                path: '/',
+                maxAge: 60 * 60 * 24, // 1 day
+                httpOnly: false,
+                secure: process.env.NODE_ENV === 'production',
+                sameSite: 'lax',
+            })
+
+            revalidatePath('/notes');
+            redirect('/notes');
+            break;
         default:
             return {
                 status: 'error',
                 message: 'Invalid action'
             } as const;
-            break;
     }
 }
